Move directory of health classes
